Fixed a couple of warnings

+++ Runtime/Graphics/D3D11/D3D11IndexBuffer.cpp
@@ -68,7 +68,7 @@ namespace Directus
 		initData.SysMemSlicePitch = 0;
 
 		// Compute memory usage
-		m_memoryUsage = sizeof(unsigned int) * indices.size();
+		m_memoryUsage = (unsigned int)(sizeof(unsigned int) * indices.size());
 
 		HRESULT result = m_graphics->GetDevice()->CreateBuffer(&bufferDesc, &initData, &m_buffer);
 		if FAILED(result)
